[2chain] add smoke test for upgrade
